Deprecate LightningModule.summarize() in favor of pl.utilities.model_summary.summarize() (#8513)
