+++ tests/accelerators/test_ddp_spawn.py
@@ -14,8 +14,8 @@
 import tests.helpers.pipelines as tpipes
 import tests.helpers.utils as tutils
 from pytorch_lightning.callbacks import EarlyStopping
-from pytorch_lightning.core import memory
 from pytorch_lightning.trainer import Trainer
+from pytorch_lightning.utilities import memory
 from tests.helpers import BoringModel
 from tests.helpers.datamodules import ClassifDataModule
 from tests.helpers.runif import RunIf
